- Modifica tipo colonna "quantity" in tabella "location_product" (da int a float); - Modifica tipo colonna "quantity_needed" e "quantity_transferred" in tabella "production_orders_products" (da int a float); - Aggiunto file sql "test_data.sql" contenente dati inviati da Luca; - Modifica DatabaseSeeder.php che ora fa il seed solo dei dati che arrivano dal file fornito; - Ottimizzazione query e risoluzione problemi di velocità nella parte di visualizzazione distinta base e dettaglio ordine di produzione; - Altri fix e miglioramenti;

+++ database/migrations/2023_05_18_122941_create_production_orders_products_table.php
@@ -15,8 +15,8 @@ return new class extends Migration
             $table->id();
 			$table->foreignIdFor(\App\Models\ProductionOrder::class, 'production_order_id');
 			$table->foreignIdFor(\App\Models\Product::class, 'product_id');
-			$table->integer('quantity_needed');
-			$table->integer('quantity_transferred');
+			$table->float('quantity_needed');
+			$table->float('quantity_transferred');
 	        $table->enum('status', array_keys(config('requested.warehouse_orders.status')));
             $table->timestamps();
         });
